# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TipoPartida": {
				"type": "string",
				"enum": [
					"debito",
					"credito"
				]
			},
			"Partida": {
				"description": "Representa uma linha de débito ou crédito de uma transação contábil.",
				"properties": {
					"id_conta": {
						"type": "number",
						"format": "double"
					},
					"tipo_partida": {
						"$ref": "#/components/schemas/TipoPartida"
					},
					"valor": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id_conta",
					"tipo_partida",
					"valor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Lancamento": {
				"description": "Representa uma transação contábil composta (com múltiplas partidas).",
				"properties": {
					"id_lancamento": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"type": "string",
						"format": "date-time"
					},
					"descricao": {
						"type": "string"
					},
					"valor_total": {
						"type": "number",
						"format": "double"
					},
					"partidas": {
						"items": {
							"$ref": "#/components/schemas/Partida"
						},
						"type": "array"
					},
					"id_empresa": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id_lancamento",
					"data",
					"descricao",
					"valor_total",
					"partidas",
					"id_empresa"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DadosTransacao": {
				"description": "Interface para a nova estrutura de dados de entrada para a criação de um lançamento (transação composta).",
				"properties": {
					"data": {
						"type": "string"
					},
					"descricao": {
						"type": "string"
					},
					"partidas": {
						"items": {
							"$ref": "#/components/schemas/Partida"
						},
						"type": "array"
					}
				},
				"required": [
					"data",
					"descricao",
					"partidas"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Empresa": {
				"properties": {
					"id_empresa": {
						"type": "number",
						"format": "double"
					},
					"razao_social": {
						"type": "string"
					},
					"cnpj": {
						"type": "string"
					}
				},
				"required": [
					"id_empresa",
					"razao_social",
					"cnpj"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TipoConta": {
				"enum": [
					"Ativo",
					"Passivo",
					"PatrimonioLiquido",
					"Receita",
					"Despesa"
				],
				"type": "string"
			},
			"Conta": {
				"properties": {
					"id_conta": {
						"type": "number",
						"format": "double"
					},
					"nome_conta": {
						"type": "string"
					},
					"tipo_conta": {
						"$ref": "#/components/schemas/TipoConta"
					},
					"subtipo_conta": {
						"type": "string"
					},
					"subtipo_secundario": {
						"type": "string"
					},
					"codigo_conta": {
						"type": "string"
					},
					"id_empresa": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id_conta",
					"nome_conta",
					"tipo_conta",
					"codigo_conta",
					"id_empresa"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial__nome_conta-string--tipo_conta-TipoConta--codigo_conta-string--subtipo_conta_63_-string--subtipo_secundario_63_-string__": {
				"properties": {
					"nome_conta": {
						"type": "string"
					},
					"tipo_conta": {
						"$ref": "#/components/schemas/TipoConta"
					},
					"codigo_conta": {
						"type": "string"
					},
					"subtipo_conta": {
						"type": "string"
					},
					"subtipo_secundario": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
...
		"/lancamentos": {
			"post": {
				"operationId": "CriarLancamento",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Lancamento"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Lançamentos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DadosTransacao"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListarLancamentos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Lancamento"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lançamentos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/lancamentos/{id}": {
			"get": {
				"operationId": "BuscarLancamentoPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Lancamento"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Lançamentos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AtualizarLancamento",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Lancamento"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Lançamentos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"partidas": {
										"items": {
											"$ref": "#/components/schemas/Partida"
										},
										"type": "array"
									},
									"descricao": {
										"type": "string"
									},
									"data": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletarLancamento",
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Lançamentos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/empresas": {
			"post": {
				"operationId": "CriarEmpresa",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Empresa"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Cria uma nova empresa no sistema.\r\nEsta rota é geralmente restrita a administradores ou ao fluxo de cadastro inicial (Sign-up).",
				"tags": [
					"Empresas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"cnpj": {
										"type": "string"
									},
									"razao_social": {
										"type": "string"
									}
								},
								"required": [
									"cnpj",
									"razao_social"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/empresas/{id}": {
			"get": {
				"operationId": "BuscarEmpresaPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Empresa"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Busca uma empresa pelo ID.",
				"tags": [
					"Empresas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/empresas/cnpj/{cnpj}": {
			"get": {
				"operationId": "BuscarEmpresaPorCnpj",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Empresa"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Busca uma empresa pelo CNPJ.",
				"tags": [
					"Empresas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cnpj",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/contas": {
			"post": {
				"operationId": "CriarConta",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Conta"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Criar Conta",
				"tags": [
					"Contas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"subtipo_secundario": {
										"type": "string"
									},
									"subtipo_conta": {
										"type": "string"
									},
									"codigo_conta": {
										"type": "string"
									},
									"id_empresa": {
										"type": "number",
										"format": "double"
									},
									"tipo_conta": {
										"$ref": "#/components/schemas/TipoConta"
									},
									"nome_conta": {
										"type": "string"
									}
								},
								"required": [
									"codigo_conta",
									"id_empresa",
									"tipo_conta",
									"nome_conta"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/contas/{id_empresa}": {
			"get": {
				"operationId": "ListarContas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Conta"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Listar contas por empresa",
				"tags": [
					"Contas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id_empresa",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/contas/{id_empresa}/{id}": {
			"get": {
				"operationId": "BuscarContaPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Conta"
										},
										{}
									]
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Buscar conta por id",
				"tags": [
					"Contas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id_empresa",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AtualizarConta",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Conta"
										},
										{}
									]
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Atualizar conta",
				"tags": [
					"Contas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id_empresa",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial__nome_conta-string--tipo_conta-TipoConta--codigo_conta-string--subtipo_conta_63_-string--subtipo_secundario_63_-string__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletarConta",
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletar conta",
				"tags": [
					"Contas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id_empresa",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}